# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  label.CreateLabelRequestBody:
    properties:
      cluster_name:
        type: string
      label_key:
        type: string
      label_value:
        type: string
      project_name:
        type: string
    required:
    - cluster_name
    - label_key
    - label_value
    - project_name
    type: object
...
  util.Res:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /label/create:
    post:
      consumes:
      - application/json
      description: Create Label
      operationId: CreateLabel
      parameters:
      - description: Create Label
        in: body
        name: ReqeustBody
        required: true
        schema:
          $ref: '#/definitions/label.CreateLabelRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: CreateLabel
      tags:
      - CreateLabel
  /label/delete:
    delete:
      consumes:
      - application/json
      description: Delete Label
      operationId: DeleteLabel
      parameters:
      - description: Project_Name
        in: query
        name: project_name
        required: true
        type: string
      - description: Cluster_Name
        in: query
        name: cluster_name
        required: true
        type: string
      - description: Label_Key
        in: query
        name: label_key
        required: true
        type: string
      - description: Label_value
        in: query
        name: label_value
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: DeleteLabel
      tags:
      - DeleteLabel
  /label/list:
    get:
      consumes:
      - application/json
      description: List Cluster's label
      operationId: ListClusterLabel
      parameters:
      - description: Project Name
        in: query
        name: project_name
        required: true
        type: string
      - description: Cluster Name
        in: query
        name: cluster_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: ListClusterLabel
      tags:
      - ListClusterLabel
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"